requirements.txt
easy-torch
easydict
packaging
setproctitle
pandas
scikit-learn
tables
sympy
openpyxl
setuptools==59.5.0
numpy==1.24.4
transformers